fetch: tweak collect
Collect functionality is going to be useful for push/gc/etc in the next PRs.
